utlra_reduced/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
google-generativeai==0.3.2
numpy==1.24.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2
requests==2.31.0
aiohttp==3.9.1